# app/Models/Picture.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Picture extends Model
{
    use HasFactory;
    protected $fillable = ['person_id','url','sort_order'];
    public function person()
    {
        return $this->belongsTo(Person::class);
    }
}